Removes modal state from store, now the state is on the sell collectibles button
