+++ src/components/erc721/collectibles/collectibles_list.tsx
@@ -8,7 +8,7 @@ import { themeBreakPoints } from '../../../themes/commons';
 import { CollectibleFilterType } from '../../../util/filterable_collectibles';
 import { CollectibleSortType } from '../../../util/sortable_collectibles';
 import { Collectible, StoreState } from '../../../util/types';
-import { Button } from '../../common/button';
+import { SellCollectiblesButton } from '../marketplace/sell_collectibles_button';
 
 import { CollectiblesCardList } from './collectibles_card_list';
 import { CollectiblesListFilter } from './collectibles_list_filter';
@@ -99,12 +99,6 @@ const Title = styled.h1`
     }
 `;
 
-const ButtonStyled = styled(Button)`
-    @media (min-width: ${themeBreakPoints.md}) {
-        margin-left: auto;
-    }
-`;
-
 export class CollectiblesList extends React.Component<Props, State> {
     public state = {
         sortType: CollectibleSortType.NewestAdded,
@@ -122,9 +116,7 @@ export class CollectiblesList extends React.Component<Props, State> {
                     <Title>{title}</Title>
                     <CollectiblesListSortStyled currentValue={sortType} onChange={this._onChangeSortType} />
                     <CollectiblesListFilterStyled currentValue={filterType} onChange={this._onChangeFilterType} />
-                    <ButtonStyled variant="quaternary" onClick={this.props.toggleCollectibleListModal}>
-                        Sell collectibles
-                    </ButtonStyled>
+                    <SellCollectiblesButton />
                 </FiltersMenu>
                 <CollectiblesCardList collectibles={collectibles} sortType={sortType} filterType={filterType} />
             </MainContainer>
