+++ src/components/erc721/erc721_app.tsx
@@ -6,7 +6,6 @@ import { AdBlockDetector } from '../common/adblock_detector';
 import { CheckMetamaskStateModalContainer } from '../common/check_metamask_state_modal_container';
 import { GeneralLayoutContainer } from '../general_layout';
 
-import { CollectibleListModal } from './collectibles/collectible_list_modal';
 import { CollectibleSellModal } from './collectibles/collectible_sell_modal';
 import { ToolbarContentContainer } from './common/toolbar_content';
 import { AllCollectibles } from './pages/all_collectibles';
@@ -19,7 +18,7 @@ export const Erc721App = () => (
     <GeneralLayoutContainer toolbar={toolbar}>
         <AdBlockDetector />
         <CollectibleSellModal />
-        <CollectibleListModal />
+
         <CheckMetamaskStateModalContainer />
         <Switch>
             <Route exact={true} path={`${ERC721_APP_BASE_PATH}/`} component={AllCollectibles} />
